requirements.txt
atari-py==0.2.9
beautifultable==0.8.0
Brotli==1.0.9
cmake==3.25.0
envpool==0.8.4
lit==15.0.7
matplotlib==3.7.5
mujoco==3.2.3
mujoco-py==2.1.2.14
opencv-python==4.10.0.84
pip==21.0
PySocks==1.7.1
tensorflow==2.13.1
torchaudio==2.0.2+cu118
torchrl==0.5.0
torchvision==0.15.2+cu118
wandb==0.18.3
